rename signaleedit to signalview
