+++ tools/cabana/detailwidget.h
@@ -8,7 +8,7 @@
 #include "tools/cabana/binaryview.h"
 #include "tools/cabana/chartswidget.h"
 #include "tools/cabana/historylog.h"
-#include "tools/cabana/signaledit.h"
+#include "tools/cabana/signalview.h"
 
 class EditMessageDialog : public QDialog {
 public:
